move pydantic tests from test_constrained_attribute_parsing
